--- src/multi_condition_comparisions/methods/__init__.py
@@ -1,1 +1,0 @@
-METHOD_REGISTRY = {}
