--- src/multi_condition_comparisions/methods/base.py
@@ -1,11 +1,0 @@
-from abc import ABC, abstractmethod
-import pandas as pd
-
-
-class BaseMethod(ABC):
-    @abstractmethod
-    @staticmethod
-    def compare_groups() -> pd.DataFrame: ...
-
-
-class LinearModelBase(BaseMethod): ...
